+++ nix/tools.nix
@@ -10,5 +10,10 @@ in
       exec ${pkgs.python3}/bin/python3 ${../tools/generate-dart-ffi-api.py} "$@"
     '';
 
+  pruneUnusedNeeded =
+    pkgs.writeShellScriptBin "flutter-haskell-prune-unused-needed" ''
+      exec ${pkgs.python3}/bin/python3 ${../tools/prune-unused-needed.py} "$@"
+    '';
+
   inherit flutterSdk androidSdk;
 }